Code cleanup for beta testing
