+++ ReboundLogParser/ReboundLogParser.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2035
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31911.196
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReboundLogParser2", "ReboundLogParser2\ReboundLogParser2.csproj", "{0AB3CBAB-452E-4795-905E-533203939E76}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReboundLogParser", "ReboundLogParser\ReboundLogParser.csproj", "{0AB3CBAB-452E-4795-905E-533203939E76}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
