+++ ReboundLogParser/ReboundLogParser/Program.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace ReboundLogParser2 {
+namespace ReboundLogParser {
   static class Program {
     /// <summary>
     /// The main entry point for the application.
@@ -13,7 +10,7 @@ namespace ReboundLogParser2 {
     static void Main() {
       Application.EnableVisualStyles();
       Application.SetCompatibleTextRenderingDefault(false);
-      Application.Run(new Form1());
+      Application.Run(new MainForm());
     }
   }
 }
